Rename `parsec/components/postgresql/queries/q_auth_lock_realm.py` -> `q_lock_realm.py` since it never do authentication!
